Date time challenge
The Portland-based company you work for just opened two new branches. One is in New York City, the other in London. They need a very simple program to find out if the branches are open or closed. The hours of both branches are 9:00 a.m.-5:00 p.m. in their own time zone.

1. Using IDLE, create a new file.

2. Import the datetime module and any others to aid in time zone calculations.

3. Create a script that will find out the current times in the Portland HQ and NYC and London branches. Then, compare that time with each branch's hours to see if they are open or closed.

4. Print out to the screen the three branches and whether they are open or closed.